+++ proxy_worker/version.py
@@ -1,1 +1,4 @@
-VERSION = "1.0.0"
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+VERSION = "4.36.1"
