+++ setup.cfg
@@ -3,7 +3,7 @@ addopts = --capture=no --assert=plain --strict --tb native
 testpaths = tests
 
 [mypy]
-python_version = 3.6
+python_version = 3.13
 check_untyped_defs = True
 warn_redundant_casts = True
 warn_unused_ignores = True
